requirements.txt
asgiref==3.3.1
Django==3.1.7
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-js-asset==1.2.2
django-modeltranslation==0.16.2
django-treasuremap==0.3.3
Pillow==8.1.2
psycopg2-binary==2.8.6
python-gettext==4.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
